DMP-4467: Performance Testing - Automated Tasks - UnstructuredToArmDataStore - Running a batch size of 50,000 you cannot see error until transaction completed or failed  (#2445)

+++ src/main/java/uk/gov/hmcts/darts/util/AsyncUtil.java
@@ -16,7 +16,7 @@ public final class AsyncUtil {
 
     }
 
-    public static void shutdownAndAwaitTermination(ExecutorService executorService, int timeout, TimeUnit unit) {
+    public static void shutdownAndAwaitTermination(ExecutorService executorService, long timeout, TimeUnit unit) {
         executorService.shutdown();
         try {
             if (!executorService.awaitTermination(timeout, unit)) {
@@ -32,7 +32,7 @@ public final class AsyncUtil {
     }
 
     public static void invokeAllAwaitTermination(List<Callable<Void>> tasks,
-                                                 int threads, int timeout, TimeUnit timeUnit) throws InterruptedException {
+                                                 int threads, long timeout, TimeUnit timeUnit) throws InterruptedException {
         log.info("Starting {} tasks with {} threads", tasks.size(), threads);
         final Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         //Add authentication to each task as auth is thread local
